before signal_types simplification

+++ ruler/src/nodes/ruler_node.cpp
@@ -125,13 +125,13 @@ void RulerNode::readParameters()
         if (type == "consumable")
         {
           resource.reset(new ruler::DiscreteConsumableResource(
-              id, name, discrete_capacity, discrete_initial_level,
+              id, name, (long) discrete_capacity, (long) discrete_initial_level,
               ros::Duration(latence)));
         }
         else
         {
           resource.reset(new ruler::DiscreteReusableResource(
-              id, name, discrete_capacity, discrete_initial_level,
+              id, name, (long) discrete_capacity, (long) discrete_initial_level,
               ros::Duration(latence)));
         }
         break;
